jinko: Rename log.rs -> debug.rs
