+++ src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod builtins;
 pub mod context;
+pub mod debug;
 pub mod error;
 #[cfg(feature = "ffi")]
 mod ffi;
@@ -10,7 +11,6 @@ mod indent;
 pub mod instance;
 pub mod instruction;
 pub mod location;
-pub mod log;
 pub mod parser;
 pub mod symbol;
 pub mod typechecker;
